Disable swap service
Do not create or enable swap at boot time.  Leave the service in place
so that swap can be enabled ad-hoc if needed.
Quite possibly this should be an option in the provisioner, but since
we've had a swapless cluster in production for a while now, it seems
unlikely that we are going to be using swap much in the future so for
now it's hard coded to be off.