Surface cloudwatch alarms in Heimdall (#223)
* Surface cloudwatch alarms in Heimdall

Heimdall supports cloudwatch alarms: https://monitoring-manager.in.ft.com/add-cloudwatch-alarm when the naming follows a convention.

This change allows us to recreate the information shown in Dashing in Heimdall, which brings us a step closer to turning off the former in favour of the latter.

* Correct systemCode label for Aurora CloudWatch

* Add TagEnvironment to aurora-cloudwatch.yml CF

Currently in aurora-cloudwatch.yml the environment of the alarm name is
equal to "p". We need to add the TagEnvironment to have the correct
environment  in the alarm name.

Jira: https://jira.ft.com/browse/CPH-111

* Update the content-test sgs

The Aurora Server SGs were outdated. One of the SGs that was configured
in the content-test.yml does not exist anymore, in its place there is
another SG.