Change how to provision elastic search domains
The point of this change is to make the elasticsearch provisioners similar to our other provisioners that do not depend on ansible. This is to lower complexity. Keep all the resources within the cloudformation. And be able to put the domain configs in the git repo and have versioning for them.